update universal-pathlib for 3.11 support
I also removed `continue-on-error`, for which I had to update `pylint` to 2.15.0.
There is `aiohttp` wheel for 3.11, so we have to also set  AIOHTTP_NO_EXTENSIONS to 1 so that we install non-cythonized build (this is only in effect when we install from sdist, otherwise if the wheels exist, they are installed instead. So this is really only going to affect Python 3.11.